Moved Util class to src and refactored.
