+++ src/ui/metadata_panel.py
@@ -7,7 +7,6 @@
 # “An Huynh” <an35@pdx.edu>
 # “Theron Anderson” <atheron@pdx.edu>
 # This software is licensed under the MIT License. See LICENSE file for the full text.
-import wx
 from src.ui.iui_behavior import IUIBehavior
 from src.ui.application_state import ApplicationState
 from src.ui.user_event import UserEvent
@@ -17,7 +16,7 @@ from src.model_conversion.model_shipper import ModelShipper
 from src.log_messages.log_message import LogMessage
 from src.log_messages.log_type import LogType
 from src.ui.ui_style import *
-from util import Util
+from src.util import Util
 
 
 class MetadataPanel(wx.Panel, IUIBehavior):
