+++ tests/ci_tests/integration/test_output_model_message.py
@@ -12,7 +12,7 @@ from src.log_messages.input_model_message import InputModelMessage
 from src.log_messages.log_type import LogType
 from src.model_conversion.ldraw_model import LDrawModel
 from src.model_conversion.model_shipper import ModelShipper
-from util import Util
+from src.util import Util
 
 
 class TestInputModelMessage(unittest.TestCase):
